+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - core, plugins: return "0x0" instead of "(nil)" for pointers formatted in strings
 - php: fix return value of function hdata_longlong
 - tcl: fix return value of function hdata_longlong ([#2119](https://github.com/weechat/weechat/issues/2119))
+- relay: fix read of one buffer line in API protocol
 - relay: fix websocket permessage-deflate extension when the client doesn't send the max window bits parameters ([#1549](https://github.com/weechat/weechat/issues/1549))
 - relay: fix allocation and reinit of field "client_context_takeover" in websocket deflate structure ([#1549](https://github.com/weechat/weechat/issues/1549))
 - core: fix completion of command `/item refresh`